refactored prep_vcfs_folder name
