--- prep_vcfs/__init__.py
@@ -1,3 +1,0 @@
-#from prep_vcfs import *
-#from prep_vcfs import *
-from prep_vcfs.utils.filters import *
