+++ prep_vcfs_somatic/__init__.py
@@ -1,3 +1,3 @@
 #from prep_vcfs_somatic import *
 #from prep_vcfs_somatic import *
-from myGenotype import Genotype
+from prep_vcfs_somatic.utils.filters import *
