+++ prep_vcfs_somatic/lancet/lancet.somatic.addFieldsForVcfMerger.py
