+++ prep_vcfs_somatic/octopus/octopus.somatic.addFieldsForVcfMerger.py
