+++ prep_vcfs_somatic/octopus/octopus.somatic.checkForBlockSubstitutionVariants.py
