+++ prep_vcfs_somatic/prep_vcf.sh
