+++ prep_vcfs_somatic/strelka2/strelka2.somatic.addFieldsForVcfMerger.py
