+++ prep_vcfs_somatic/utils/convert_contig_list_from_bam_to_vcf_format.sh
