+++ prep_vcfs_somatic/utils/filter_vcf.sh
