+++ prep_vcfs_somatic/vardictjava/vardictjava.somatic.addFieldsForVcfMerger.py
